typecast size of char array to int

+++ 0x02-functions_nested_loops/0-putchar.c
@@ -13,7 +13,7 @@ int main(void)
 	char word[8] = "_putchar";
 	int i;
 
-	for (i = 0; i < sizeof(word); i++)
+	for (i = 0; i < (int) sizeof(word); i++)
 	{
 		putchar(word[i]);
 	}
